requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemyPlus==0.7.4
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.1.4
Werkzeug==0.11.11
Whoosh==2.7.4
blinker==1.4
click==6.6
itsdangerous==0.24
lxml==3.7.0
requests==2.12.3
